backend/requirements.txt
pandas
python-docx
openpyxl
pdfplumber==0.11.9
pypdf==6.7.1
pdfminer.six
pypdfium2
supabase
python-dotenv
playwright==1.58.0
easyocr
torch
torchvision
opencv-python-headless
imutils
numpy
requests
beautifulsoup4
